requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.14
